fetching user detail with userid

+++ src/main/java/com/attendanceapp/model/UserService.java
@@ -3,9 +3,9 @@ package com.attendanceapp.model;
 public interface UserService
 {
     public void insertUser(User u );
-    public  User getUserById(int userid);
-    public void updateUser(int userid);
-    public void deleteUser(int userid);
+    public  User getUserById(long userid);
+    public void updateUser(long userid);
+    public void deleteUser(long userid);
 
 
 }
